fix(knative): address review — SDK-compatible tests, caller coverage, presence check
- Extract buildAutoscalingPatch to autoscalingPatch.ts so tests import
  the pure function without the SDK module resolution quirk.
- Add caller-level tests that simulate the form→builder boundary for
  annotation clearing, containerConcurrency clearing, absent-field
  no-op, and empty-string-value presence edge case.
- Fix presence predicate: use 'key' in prevAnns instead of truthiness
  so an existing annotation with empty-string value is correctly
  treated as present (deletion, not no-op).
- All 35 tests pass under npm run test (headlamp-plugin test).
