+++ knative/src/resources/knative/kservice.ts
@@ -19,7 +19,14 @@ import {
   KubeObject,
   type KubeObjectInterface,
 } from '@kinvolk/headlamp-plugin/lib/k8s/cluster';
+import {
+  type AutoscalingParams,
+  type AutoscalingPatchBody,
+  buildAutoscalingPatch,
+} from './autoscalingPatch';
 import type { Condition } from './common';
+
+export type { AutoscalingParams, AutoscalingPatchBody };
 
 export type Traffic = {
   percent: number;
@@ -28,33 +35,6 @@ export type Traffic = {
   configurationName?: string;
   tag?: string;
   url?: string;
-};
-
-export type AutoscalingParams = {
-  metric?: 'concurrency' | 'rps' | null;
-  target?: number | null;
-  targetUtilization?: number | null;
-  containerConcurrency?: number | null;
-  minScale?: number | null;
-  maxScale?: number | null;
-  initialScale?: number | null;
-  activationScale?: number | null;
-  scaleDownDelay?: string | null;
-  stableWindow?: string | null;
-};
-
-export type AutoscalingPatchBody = {
-  spec: {
-    template: {
-      metadata?: {
-        name?: string | null;
-        annotations?: Record<string, string | null>;
-      };
-      spec?: {
-        containerConcurrency?: number | null;
-      };
-    };
-  };
 };
 
 interface KServiceResource extends KubeObjectInterface {
@@ -146,62 +126,7 @@ export class KService extends KubeObject<KServiceResource> {
   }
 
   static buildAutoscalingPatch(params: AutoscalingParams): AutoscalingPatchBody | null {
-    const {
-      metric,
-      target,
-      targetUtilization,
-      minScale,
-      maxScale,
-      initialScale,
-      activationScale,
-      scaleDownDelay,
-      stableWindow,
-      containerConcurrency,
-    } = params;
-
-    const annotationSources: Record<string, string | number | null | undefined> = {
-      'autoscaling.knative.dev/metric': metric,
-      'autoscaling.knative.dev/target': target,
-      'autoscaling.knative.dev/target-utilization-percentage': targetUtilization,
-      'autoscaling.knative.dev/min-scale': minScale,
-      'autoscaling.knative.dev/max-scale': maxScale,
-      'autoscaling.knative.dev/initial-scale': initialScale,
-      'autoscaling.knative.dev/activation-scale': activationScale,
-      'autoscaling.knative.dev/scale-down-delay': scaleDownDelay,
-      'autoscaling.knative.dev/window': stableWindow,
-    };
-
-    const annotationsPatch: Record<string, string | null> = {};
-    for (const [key, value] of Object.entries(annotationSources)) {
-      if (typeof value === 'undefined') {
-        continue;
-      }
-      annotationsPatch[key] = value === null ? null : String(value);
-    }
-
-    const templateSpecPatch: { containerConcurrency?: number | null } = {};
-    if (typeof containerConcurrency !== 'undefined') {
-      templateSpecPatch.containerConcurrency = containerConcurrency;
-    }
-
-    const hasAnnotationsPatch = Object.keys(annotationsPatch).length > 0;
-    const hasTemplateSpecPatch = Object.keys(templateSpecPatch).length > 0;
-
-    if (!hasAnnotationsPatch && !hasTemplateSpecPatch) {
-      return null;
-    }
-
-    return {
-      spec: {
-        template: {
-          metadata: {
-            name: null as any,
-            ...(hasAnnotationsPatch ? { annotations: annotationsPatch } : {}),
-          },
-          ...(hasTemplateSpecPatch ? { spec: templateSpecPatch } : {}),
-        },
-      },
-    };
+    return buildAutoscalingPatch(params);
   }
 
   static getBaseObject(): KServiceResource {
